refactor(ds4): modularize ds4 code
